Laborator 3

+++ M_Sinca_Teodora_Ioana_Lab2/Controllers/BooksController.cs
@@ -20,14 +20,46 @@ namespace M_Sinca_Teodora_Ioana_Lab2.Controllers
         }
 
         // GET: Books
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string sortOrder, string searchString)
         {
-            var books = await _context.Book
-               .Include(b => b.Author)
-               .Include(b => b.Genre)
-               .ToListAsync();
+            ViewData["TitleSortParm"] = String.IsNullOrEmpty(sortOrder) ? "title_desc" : "";
+            ViewData["PriceSortParm"] = sortOrder == "Price" ? "price_desc" : "Price";
+            ViewData["CurrentFilter"] = searchString;
+            var books = from b in _context.Book
+                        join a in _context.Author on b.AuthorID equals a.ID
+                        select new BookViewModel
+                        {
+                            ID = b.ID,
+                            Title = b.Title,
+                            Price = b.Price,
+                            FullName = a.FullName
+                        };
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                books = books.Where(s => s.Title.Contains(searchString));
+            }
+            switch (sortOrder)
+            {
+                case "title_desc":
+                    books = books.OrderByDescending(b => b.Title);
+                    break;
+                case "Price":
+                    books = books.OrderBy(b => b.Price);
+                    break;
+                case "price_desc":
+                    books = books.OrderByDescending(b => b.Price);
+                    break;
+                default:
+                    books = books.OrderBy(b => b.Title);
+                    break;
+            }
+            return View(await books.AsNoTracking().ToListAsync());
+           /*  var books = await _context.Book
+                 .Include(b => b.Author)
+                 .Include(b => b.Genre)
+                 .ToListAsync();
 
-            return View(books);
+              return View(books);*/
         }
 
         // GET: Books/Details/5
@@ -38,7 +70,7 @@ namespace M_Sinca_Teodora_Ioana_Lab2.Controllers
                 return NotFound();
             }
 
-            var book = await _context.Book.Include(b => b.Genre).Include(b => b.Author).FirstOrDefaultAsync(m => m.ID == id);
+            var book = await _context.Book.Include(b => b.Genre).Include(b => b.Author).Include(s => s.Orders).ThenInclude(e => e.Customer).FirstOrDefaultAsync(m => m.ID == id);
             if (book == null)
             {
                 return NotFound();
@@ -68,6 +100,8 @@ namespace M_Sinca_Teodora_Ioana_Lab2.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ID,Title,AuthorID,Price,GenreID")] Book book)
         {
+            try
+            { 
             if (ModelState.IsValid)
             {
                 _context.Add(book);
@@ -84,6 +118,13 @@ namespace M_Sinca_Teodora_Ioana_Lab2.Controllers
                 "ID",
                 "FullName"
             );
+            }
+            catch (DbUpdateException /* ex*/)
+            {
+
+                ModelState.AddModelError("", "Unable to save changes. " +
+                "Try again, and if the problem persists ");
+            }
             return View(book);
         }
 
@@ -125,6 +166,24 @@ namespace M_Sinca_Teodora_Ioana_Lab2.Controllers
                 return NotFound();
             }
 
+            var bookToUpdate = await _context.Book.FirstOrDefaultAsync(s => s.ID == id);
+            if (await TryUpdateModelAsync<Book>(
+             bookToUpdate,
+             "",
+             s => s.AuthorID, s => s.Title, s => s.Price))
+            {
+                try
+                {
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                catch (DbUpdateException /* ex */)
+                {
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists");
+                }
+            }
+
             if (ModelState.IsValid)
             {
                 try
@@ -146,20 +205,24 @@ namespace M_Sinca_Teodora_Ioana_Lab2.Controllers
                 return RedirectToAction(nameof(Index));
             }
             ViewData["GenreID"] = new SelectList(_context.Set<Genre>(), "ID", "Name", book.GenreID);
-            ViewData["AuthorID"] = new SelectList(
-                 _context.Set<Author>().Select(a => new
-                 {
-                     ID = a.ID,
-                     FullName = a.FirstName + " " + a.LastName
-                 }),
-                 "ID",
-                 "FullName"
-             );
-            return View(book);
+            //ViewData["AuthorID"] = new SelectList(
+            // _context.Set<Author>().Select(a => new
+            //{
+            //  ID = a.ID,
+            //FullName = a.FirstName + " " + a.LastName
+            //}),
+            //"ID",
+            //"FullName"
+            //);
+
+            //return View(book);
+            ViewData["AuthorID"] = new SelectList(_context.Author, "ID", "FullName",
+bookToUpdate.AuthorID);
+            return View(bookToUpdate);
         }
 
         // GET: Books/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null || _context.Book == null)
             {
@@ -167,13 +230,17 @@ namespace M_Sinca_Teodora_Ioana_Lab2.Controllers
             }
 
             var book = await _context.Book
-                .Include(b => b.Genre).Include(b => b.Author)
+                .Include(b => b.Genre).Include(b => b.Author).AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (book == null)
             {
                 return NotFound();
             }
-
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                "Delete failed. Try again";
+            }
             return View(book);
         }
 
@@ -187,13 +254,27 @@ namespace M_Sinca_Teodora_Ioana_Lab2.Controllers
                 return Problem("Entity set 'M_Sinca_Teodora_Ioana_Lab2Context.Book'  is null.");
             }
             var book = await _context.Book.FindAsync(id);
+            if (book == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            /*
             if (book != null)
             {
                 _context.Book.Remove(book);
             }
-            
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            */
+            try
+            {
+                _context.Book.Remove(book);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool BookExists(int id)
